refactor: improvement structure of id card folder
