+++ src/components/IdCard/index.tsx
@@ -1,11 +1,9 @@
-import { useEffect, useRef, useState } from 'react';
-import html2canvas from 'html2canvas';
-import { Typography } from '../Typography';
-import * as S from './styles';
 import { Button } from '@chakra-ui/react';
 import { useTheme } from 'styled-components';
-import { capitalize } from '../../utils/capitalize';
-import { getKeyByCourse } from './utils';
+import { capitalize } from '../../utils';
+import { Typography } from '../Typography';
+import { useIdCardHandlers } from './idCardHandlers';
+import * as S from './styles';
 
 export type IdCardProps = {
   name?: string;
@@ -32,48 +30,12 @@ export const IdCard = ({
     alt: 'imagem aluno(a)',
   },
 }: IdCardProps) => {
-  const [imageSrc, setimageSrc] = useState<{
-    src: string;
-    alt: string;
-  }>(image);
-  const ref = useRef<HTMLDivElement>(null);
   const theme = useTheme();
-  const courseKey = getKeyByCourse(course);
 
-  useEffect(() => {
-    const imageUploaded = JSON.parse(
-      sessionStorage.getItem('id-card-image') || '{}'
-    );
-
-    if (imageUploaded) {
-      setimageSrc((prev) => ({ ...prev, ...imageUploaded.image }));
-    }
-  }, []);
-
-  const handleDownloadImage = async () => {
-    const element = ref?.current;
-
-    if (!element) {
-      return;
-    }
-    const canvas = await html2canvas(element, {
-      backgroundColor: null,
-      allowTaint: true,
-    });
-    const data = canvas.toDataURL('image/jpg');
-    const link = document.createElement('a');
-
-    if (typeof link.download === 'string') {
-      link.href = data;
-      link.download = 'image.jpg';
-
-      document.body.appendChild(link);
-      link.click();
-      document.body.removeChild(link);
-    } else {
-      window.open(data);
-    }
-  };
+  const { imageSrc, courseKey, ref, handleDownloadImage } = useIdCardHandlers({
+    image,
+    course,
+  });
 
   return (
     <>
